Prevent students from committing config...
This automatically runs a 'git' command which will prevent
the config.json file from being detected with modifications.

The magic spell to reverse this action is also included in the comment.

The command was specifically generalized to work regardless of the
starting directory from where it was called. Thanks for the getPath()
function which did all the hard work to make this possible.